requirements.txt
coverage==7.3.0
django-ninja==1.2.2
django-ninja-extra==0.21.1
django-ninja-jwt==5.3.2
ninja-schema==0.13.6
Django==5.0.6
environ==1.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytest-mock==3.11.1
pytest==7.4.0
requests==2.31.0
pydantic==2.7.4